pyproject.toml
[tool.poetry]
name = "hw-12"
version = "0.1.0"
description = ""
authors = ["chabanchuk <145770156+chabanchuk@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
uvicorn = "^0.29.0"
pydantic = "^1.10.8"
email-validator = "^1.3.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
